fix(seo): add glossary metadata on live /education/glossary URL.
Move DefinedTermSet JSON-LD and page metadata to the canonical route;
strip duplicate SEO from the legacy /learn/glossary layout.
